# Performance: Use one 24 byte random buffer instead of calling randomBytes() twice